docker_compose/simple_project/app/requirements.txt
Django==5.0.4
psycopg[binary]==3.2.3
python-dotenv==1.0.1
split-settings==1.0.0
uwsgi==2.0.28
Pillow
django-cors-headers
django-debug-toolbar